requirements-from-project.txt
mmcv
pyrallis
ipdb
accelerate
beautifulsoup4
bs4
diffusers
clip
timm
xformers
came-pytorch
ftfy
image-reward
omegaconf
opencv-python
optimum
patch_conv
peft
protobuf
pytorch-fid
regex
sentencepiece
triton
wandb
webdataset
yapf